requirements.txt
asgiref==3.3.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
